Adding the model and the data collection files
The model contains all the files that *should* train the model while data collection opens up an instance of the game and allows you to play it while saving the actions taken and the observations of the game state to created directories called 'actions' and 'observations'. Run this file first - it generates the files that model.py relies on to construct a DataSet object to train on.